feat: :sparkles: Adds string type in policies and update documentation.
The use of ImplementationPolicy allows to customize all the container types except the key_type which is std::basic_string. This type does not handle custom pointer types such as boost::interprocess::offset_ptr. Moving std::basic_string into the ImplementationPolicy scope does not change the default behavior of the library but allow the issue.

The documentation now describe key_type as a ContiguousContainer instead
of std::basic_string.